Refactoring of communication (CommunicationService split into -> WebSocketService and HttpService)

+++ Client/Unity Project/Assets/Scripts/Communication/CommandDTOs/BossStatusUpdateCommandDTO.cs
@@ -20,8 +20,8 @@ namespace Assets.Scripts.Communication.CommandDTOs
             public class BossStatusUpdateParameters
             {
                 public String bossName = "";
-                public Int64 currentBossLife = 10000;
-                public Int64 constantBossLife = 10000;
+                public Int32 currentBossLife = 10000;
+                public Int32 constantBossLife = 10000;
                 public String status = "ALIVE";
             }
         }
